Created a change log
I think its important that the framework inform users of what the changes to the framework are, as well as being able to communicate backwards compatibility breaks so they know the impact if they can freely upgrade, or if they have some refactoring ahead of them.

With this change log, I think its also important that you begin to `tag` your releases, so people can downgrade or pick a particular version to upgrade to in case they cannot upgrade to the master branch, but can upgrade to an earlier version. Of course that would be a manual thing for them as the updater does not support choosing a version, but its still pretty good so people have options as the framework matures.